+++ utilities/js/libs/custom-select/custom-select.php
@@ -4,19 +4,22 @@ define( 'CUSTOM_SELECT_DIRECTORY', dirname( __FILE__ ) );
 define( 'CUSTOM_SELECT_DIRECTORY_URI', UTILITIES_DIRECTORY_URI . '/js/libs/custom-select' );
 
 // register custom-select lib
-add_action('wp_enqueue_scripts', function() {
+add_action( 'wp_enqueue_scripts', function() {
     wp_register_script('custom-select-module', CUSTOM_SELECT_DIRECTORY_URI . '/module/custom-select.min.js', [], '1.1.15', true);
     wp_register_script('custom-select', CUSTOM_SELECT_DIRECTORY_URI . '/app.js', ['custom-select-module'], false, true);
 
-    $selectors = get_option('custom-select-css-selectors') ?: '';
-    wp_localize_script('custom-select', 'wp_localize_script_vars', [
-        'custom_selectors' => $selectors,
-    ]);
+    wp_localize_script('custom-select', 'customSelectVars', [
+        'selectors' => trim( get_option( 'custom-select-css-selectors', '' ) ),
+    ] );
 
     if (get_option('custom-select')) {
         wp_enqueue_script('custom-select');
     }
-});
+} );
+
+add_action( 'customize_controls_enqueue_scripts', function() {
+	wp_enqueue_script( 'custom-select-be', CUSTOM_SELECT_DIRECTORY_URI . '/admin.js', [], false, true );
+} );
 
 // enable/disable custom-select customizer setting
 add_action( 'customize_register', function( $wp_customize ) {
@@ -26,8 +29,8 @@ add_action( 'customize_register', function( $wp_customize ) {
     ) );
 
     $wp_customize->add_control( 'custom-select', array(
-        'label' => __( 'Enable custom-select', 'utilities' ),
-        'description' => __( "Replacing and enhancing the native HTML select element.", 'utilities' ),
+        'label' => __( 'Enable custom-select.js', 'utilities' ),
+        'description' => __( "Replacing and enhancing the native <code>select</code> (dropdown) element.", 'utilities' ),
         'section' => 'utilities',
         'type' => 'checkbox'
     ) );
@@ -38,10 +41,10 @@ add_action( 'customize_register', function( $wp_customize ) {
     ) );
 
     $wp_customize->add_control( 'custom-select-css-selectors', array(
-        'label'       => __( 'CSS Selectors for Custom Select Field', 'utilities' ),
-        'description' => __( "Specify the CSS selectors of the select fields you want to customize. Leave this field empty to apply the custom styling to all select fields.", 'utilities' ),
+//        'label'       => __( 'CSS Selectors', 'utilities' ),
+        'description' => __( "CSS selectors for <code>select</code> fields to apply custom-select.js to. Leave blank to apply to all. Each selector in a new line.", 'utilities' ),
         'section'     => 'utilities',
         'type'        => 'textarea'
     ) );
 
-}, 9 );
+} );